requirements.txt
# Deep Research Agent - Dependencies

# Core LLM
google-generativeai>=0.4.0

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0

# HTTP client
httpx>=0.26.0

# Data validation
pydantic>=2.5.3

# Environment management
python-dotenv>=1.0.0

# Retry logic
tenacity>=8.2.3

# Search (optional)
duckduckgo-search>=4.1.0

# Vector storage (optional, for enhanced memory)
# chromadb>=0.4.22

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
